Big changes
Update the FSWeb icon to display properly on safari. Change the slider3 image. Change the text in the AboutSection. Change the Adress in the Footer and Contact Page. Centered the links in the Navbar. Added a destintion to one ServiceCard. Added 2 buttons in the ServiceDetail. Added a new Whatsapp component to every page.